+++ core/modules/cart/store/actions/couponActions.ts
@@ -1,0 +1,26 @@
+import { CartService } from '@vue-storefront/core/data-resolver'
+
+const couponActions = {
+  async removeCoupon ({ getters, dispatch }) {
+    if (getters.isTotalsSyncEnabled && getters.isCartConnected) {
+      const { result } = await CartService.removeCoupon()
+
+      if (result) {
+        dispatch('syncTotals', { forceServerSync: true })
+        return result
+      }
+    }
+  },
+  async applyCoupon ({ getters, dispatch }, couponCode) {
+    if (couponCode && getters.isTotalsSyncEnabled && getters.isCartConnected) {
+      const { result } = await CartService.applyCoupon(couponCode)
+
+      if (result) {
+        dispatch('syncTotals', { forceServerSync: true })
+      }
+      return result
+    }
+  }
+}
+
+export default couponActions
